requirements/production.txt
Django==1.9.4
Unipath==1.1
dj-database-url==0.4.0
django-debug-toolbar==1.4
psycopg2==2.6.1
pyparsing==2.2.0
python-decouple==3.0
pytz==2016.10
wsgiref==0.1.2
xlrd==0.9.4
django-widget-tweaks==1.4.1
setproctitle==1.1.9
gunicorn==19.4.5
colour==0.1.2
jira==1.0.10
